feat: :sparkles: I implemented the Layout component and integrated it in the app
I created the `Layout` component to structure the application.
It includes a header, an aside with navigation links (Todos, Add Todo, and Profiles), and a section for rendering child components.
React Icons were used for navigation links. Additionally,
I wrapped the application in the `Layout` component in `_app.js` to ensure a consistent layout across all pages.