feat: :sparkles: I added the User model with Mongoose
I installed Mongoose and defined the `User` model in the `models/User.js` file.
The model includes fields for `email`, `password`, `name`, `lastName`, `todos` (as an array of objects with `title` and `status`), and a `createdAt` field with a default value and immutability.
This model will be used to manage user data in the database.