feat: :sparkles: I created the signup form and integrated it in the signup page
I created the `SignupPage` component in `components/template/SignupPage.js` to render a registration form.
The form includes fields for email and password, as well as a register button and a link to the signin page for existing users. Then, I integrated this form in the `pages/signup.js` file to render it as the signup page.