feat: :sparkles: I added session handling to restrict access for authenticated users
I updated the login and signup forms to restrict access for authenticated users. By using `useSession` from NextAuth and `useEffect`, users who are already logged in are redirected to the homepage.
Additionally, I wrapped the entire application in the `SessionProvider` in `_app.js` to enable session management across the app.