feat: :sparkles: I created the AddTodo page and component
I developed the `AddTodoPage` component in components/template/AddTodoPage.js` to render a form for adding new todos.
The form includes a text input for the title and a radio button for selecting the status (defaulting to "todo").
Additionally, I integrated the component in the `pages/add-todo.js` file to render the Add Todo page.