feat: :sparkles: I added API integration and notifications to AddTodoPage
I updated the `AddTodoPage` component to integrate with the `/api/todos` endpoint.
The `addHandler` function sends a POST request to save the new todo and resets the form upon success.
Additionally, I installed `react-toastify` and implemented toast notifications to provide feedback to the user when a todo is added successfully.