feat: :sparkles: I added an API endpoint for updating todo status
I updated the `pages/api/todos.js` API to include a `PATCH` method for updating the status of a todo.
The API validates the request data, retrieves the user from the database, and uses the `$set` operator to update the status of the specified todo by its ID.
Appropriate status codes and responses are returned based on the outcome of the operation.