feat: :sparkles: I added functionality to update todo status in Task component
I updated the `Task` component to include functionality for updating the status of todos.
A `changeStatus` function was implemented to send a `PATCH` request to the `/api/todos` endpoint with the todo ID and the new status.
The `fetchTodos` function is called after a successful update to refresh the data.
"Back" and "Next" buttons now trigger the status update dynamically based on the provided props.