feat: :sparkles: I created ProfileForm and integrated it into ProfilePage
I developed the `ProfileForm` component in `components/module/ProfileForm.js` to render a form for updating user profile information, including fields for `name`, `lastName`, and `password`.
The component accepts state variables and a `submitHandler` as props for flexibility.
I then integrated the form into the `ProfilePage` component in `components/template/ProfilePage.js` and implemented initial state handling for the input fields. Finally, the `ProfilePage` was used in the `pages/profile.js` file to render the profile page.